feat: implement homepage shell with deferred 3D loading
